[PC-425] refactor: infra:s3 모듈 생성
